# src/main/resources/static/swagger.yml
openapi: 3.0.0
info:
  title: PayForLegalAid API
  version: 1.0.0
  description: >
    API for accessing financial reports related to legal aid payments.
    **Authentication Notice**: 
    All endpoints are secured using OAuth 2.0 with Azure Graph. A valid access token must be provided 
    in the `Authorization` header of each request in the following format: 
    `Authorization: Bearer <ACCESS_TOKEN>`.

  contact:
    name: Ministry of Justice

components:
  links:
    ReportDetailsLink:
      operationId: getReportById
      description: >
        Link to retrieve detailed metadata for a specific report using its ID.
      parameters:
        id: $response.body#/reportList/{id}
    CsvDownloadLink:
      operationId: downloadCsv
      description: Download the report as a CSV stream using its ID.
      parameters:
        id: $response.body#/id
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    UnauthorizedError:
      description: Unauthorized. A valid OAuth2 token is required.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized access. OAuth2 token required."
    NotFoundError:
      description: Resource not found in the database.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "ReportIdNotFoundException: Resource not found."
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "InternalServerError: An unknown error occurred."
    BadRequestError:
      description: The request is invalid or improperly formatted.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "BadRequestException: The request could not be processed."

security:
  - BearerAuth: [ ]

paths:
  /reports:
    get:
      tags:
        - Reports
      summary: List All Available Reports
      description: >
        Retrieves a list of all available reports, including each report's ID, name, metadata, 
        and a link to retrieve detailed metadata for each report.
        **Note**: The `reportDetailUrl` field in each report provides a direct link to `/report/{id}`.
      responses:
        '200':
          description: A JSON array containing metadata for each report.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportList:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        reportName:
                          type: string
                          example: "Invoice_analysis"
                        description:
                          type: string
                          example: "Summary of invoices by source system"
          links:
            reportDetails:
              $ref: '#/components/links/ReportDetailsLink'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /reports/{id}:
    get:
      operationId: getReportById
      tags:
        - Reports
      summary: Retrieve Report Metadata by ID
      description: Fetches metadata for a specific report using the provided report ID. This includes the report name, ID, and download URL.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
          description: The unique ID of the requested report.
      responses:
        '200':
          description: JSON object with metadata for the specified report.
          content:
            application/json:
              examples:
                Report1:
                  summary: Example Retrieve Report Metadata by ID
                  value:
                    id: 1
                    reportName: "Monthly Revenue Report"
                    reportDownloadUrl: "https://example.com/reports/1"
                Report2:
                  summary: Example  Retrieve Report Metadata by ID 2
                  value:
                    id: 2
                    reportName: "Quarterly Profit Analysis"
                    reportDownloadUrl: "https://example.com/reports/2"
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  reportName:
                    type: string
                    example: "Invoice_analysis-analysis"
                  reportDownloadUrl:
                    type: string
                    format: uri
                    example: "https://<service_destination>/csv/1"
          links:
            downloadCsv:
              $ref: '#/components/links/CsvDownloadLink'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /csv/{id}:
    get:
      tags:
        - Reports
      summary: Download Report as CSV Stream
      description: Provides a CSV data stream for a specified report ID. This can trigger a file download when accessed via a browser.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
          description: The unique ID of the requested report.
      responses:
        '200':
          description: A CSV data stream of the report. The data is streamed as an attachment.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                example: |
                  "date","amount","description"
                  "2023-01-01","1000","Legal Aid Case Payment"
          headers:
            Content-Disposition:
              description: Specifies that the response is a downloadable attachment with a file name.
              schema:
                type: string
                example: "attachment; filename=report.csv"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'